blend between kaleido effect and regular obj view

+++ src/Games/Kaleido/Kaleido.h
@@ -37,5 +37,5 @@ private:
     
     map<int,Shape*> shapes;
     
-    int             countDown;
+    float           countDown;
 };
